Use the most frequent pinyin transcription for single characters
(instead of the 1st one in alphabetical order).  This is done by
embedding a new database, which is an excerpt from the Unihan database
that just contains Pinyin transcriptions for each character (about 150
Ko compressed, 500Ko uncompressed).

This should solve the single biggest issue users might have with this
add-on.