Fix `has_any_field` and `has_field`
`has_field` expects a list of field names, but `has_any_field` iterates
over a list of field names and passes them to `has_field`, which causes
`has_any_field` to iterate over characters in a string instead of
strings in a list. Fix `has_any_field` to always pass lists to
`has_field`, and add a precondition check to `has_field` to catch this
issue in the future.